pyproject.toml
[tool.poetry]
name = "weather-zone"
version = "0.1.0"
license = "MIT"
description = "Weather forecast application"
authors = ["Ian Buttimer <ibuttimer@yahoo.com>"]
readme = "README.md"
packages = [{include = "weather_zone"}]

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.3"
django-environ = "^0.10.0"
googlemaps = "^4.10.0"
beautifulsoup4 = "^4.12.2"
lxml = "^4.9.3"
django-countries = "^7.5.1"
xmltodict = "^0.13.0"


[tool.poetry.group.dev.dependencies]
pylint = "^2.17.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
